Add the needed methods to the AuthenticationService core
