+++ demo-shell/src/app/services/index.ts
@@ -16,5 +16,4 @@
  */
 
 export * from './auth-bearer.interceptor';
-export * from './authentication.service';
 export * from './auth-token-processor.service';
